Split gamma correction into separate file.
